Add Rcpp implementation of the core Ipfp algorithm
Instead of relying on aperm like permutation of the seed as the original Ipfp-function, this implementation calculates indices for all groups of values defined by each target margin. It then loops over all targets and over all groups within these targets, summing, calculating the update factor, and applying it. This sets up the function for easy parallelization later on.

NA values in a target margin lead to the indices of the respective group being removed at the beginning, thus leading to no additional comparison or calculations during the loop.