Fix bug when NA in first margin
Removing indices for targets with NA implies that not every element of the vector will be updated when updating a margin with NA in the target. Therefore, when updating the result vector from elements in result_tmp for t==0 not all values in result are replaced if the first target margin contains NA targets.
This commit therefore performs a deep copy instead of a swap for result from result_tmp at the beginning of each iteration.